Monkey patch yum pid file to be outside of /var/run/
Yum pid file (aka lock) is hardcoded to /var/run/yum.pid.
This causes problems on newer distros where /var/run is supposed
to be a symlink to /run instead of a folder.
We work around this by running yum via python instead of as a subprocess
and monkey patch the yum module to change the location of the pid file.